refactor: restructure workflows directory
... enables workflows from anywhere in the backend modules to be registered
