+++ src/apps/backend/modules/workflow/workflow_service.py
@@ -15,7 +15,7 @@ from modules.workflow.errors import (
     WorkflowStartError,
 )
 from modules.workflow.types import QueueWorkflowParams, SearchWorkflowByIdParams
-from workflows import WORKFLOW_MAP
+from workflows.workflow_registry import WORKFLOW_MAP
 
 
 class WorkflowService:
@@ -26,25 +26,19 @@ class WorkflowService:
 
     @staticmethod
     async def _get_temporal_workflow_details(params: SearchWorkflowByIdParams) -> dict:
-        client = await Client.connect(
-            ConfigService.get_string("TEMPORAL_SERVER_ADDRESS")
-        )
+        client = await Client.connect(ConfigService.get_string("TEMPORAL_SERVER_ADDRESS"))
 
         runs = []
 
         async for info in client.list_workflows(f"WorkflowId = '{params.id}'"):
-            handle = client.get_workflow_handle(
-                workflow_id=params.id, run_id=info.run_id
-            )
+            handle = client.get_workflow_handle(workflow_id=params.id, run_id=info.run_id)
             info = await handle.describe()
 
             result = None
             if info.status and info.status.name == "COMPLETED":
                 history = await handle.fetch_history()
                 result_event = history.events[-1]
-                result_data = result_event.workflow_execution_completed_event_attributes.result.payloads[
-                    0
-                ].data
+                result_data = result_event.workflow_execution_completed_event_attributes.result.payloads[0].data
                 result = result_data.decode("utf-8")
 
             runs.append(
@@ -71,9 +65,7 @@ class WorkflowService:
 
     @staticmethod
     async def _queue_temporal_workflow(params: QueueWorkflowParams) -> str:
-        client = await Client.connect(
-            ConfigService.get_string("TEMPORAL_SERVER_ADDRESS")
-        )
+        client = await Client.connect(ConfigService.get_string("TEMPORAL_SERVER_ADDRESS"))
 
         if params.name not in WORKFLOW_MAP.keys():
             raise WorkflowNameNotFoundError(workflow_name=params.name)
@@ -95,9 +87,7 @@ class WorkflowService:
 
     @staticmethod
     async def _cancel_temporal_workflow(params: SearchWorkflowByIdParams) -> None:
-        client = await Client.connect(
-            ConfigService.get_string("TEMPORAL_SERVER_ADDRESS")
-        )
+        client = await Client.connect(ConfigService.get_string("TEMPORAL_SERVER_ADDRESS"))
         handle = client.get_workflow_handle(params.id)
 
         if await WorkflowService._get_temporal_workflow_status(handle) == "COMPLETED":
@@ -113,9 +103,7 @@ class WorkflowService:
 
     @staticmethod
     async def _terminate_temporal_workflow(params: SearchWorkflowByIdParams) -> None:
-        client = await Client.connect(
-            ConfigService.get_string("TEMPORAL_SERVER_ADDRESS")
-        )
+        client = await Client.connect(ConfigService.get_string("TEMPORAL_SERVER_ADDRESS"))
         handle = client.get_workflow_handle(params.id)
 
         if await WorkflowService._get_temporal_workflow_status(handle) == "COMPLETED":
@@ -129,9 +117,7 @@ class WorkflowService:
     @staticmethod
     def get_workflow_details(*, params: SearchWorkflowByIdParams) -> dict:
         try:
-            res = asyncio.run(
-                WorkflowService._get_temporal_workflow_details(params=params)
-            )
+            res = asyncio.run(WorkflowService._get_temporal_workflow_details(params=params))
 
         except RPCError:
             raise WorkflowIdNotFoundError(workflow_id=params.id)
@@ -143,21 +129,14 @@ class WorkflowService:
         workflows = []
 
         for name, details in WORKFLOW_MAP.items():
-            workflows.append(
-                {
-                    "name": name,
-                    "priority": details["priority"],
-                }
-            )
+            workflows.append({"name": name, "priority": details["priority"]})
 
         return workflows
 
     @staticmethod
     def queue_workflow(*, params: QueueWorkflowParams) -> str:
         try:
-            workflow_id = asyncio.run(
-                WorkflowService._queue_temporal_workflow(params=params)
-            )
+            workflow_id = asyncio.run(WorkflowService._queue_temporal_workflow(params=params))
 
         except RPCError:
             raise WorkflowStartError(workflow_name=params.name)
